+++ src/ast/visit/inventory.rs
@@ -2874,7 +2874,8 @@ macro_rules! ast_traversal {
         pub fn walk_connection_usage_member<V: $Visitor>(visitor: &mut V, node: &$($mutability)? Node<ConnectionUsageMember>) {
             visitor.enter_node(&$($mutability)? node.span);
             visitor.visit_span(&$($mutability)? node.span);
-            let ConnectionUsageMember { name, type_reference, multiplicity, connect_from, connect_to, connect_extra_ends, body, subsets, redefines, membership, by_reference } = &$($mutability)? node.value;
+            let ConnectionUsageMember { is_abstract, name, type_reference, multiplicity, connect_from, connect_to, connect_extra_ends, body, subsets, redefines, membership, by_reference } = &$($mutability)? node.value;
+            let _ = is_abstract;
             if let Some(inner) = name {
                 visitor.visit_declaration_name(inner);
             }